Create draggable file item component
