+++ client/DAW/components/index.js
@@ -6,3 +6,4 @@ export { default as Timeline } from './Timeline';
 export { default as Track } from './track';
 export { default as TrackControls } from './track-controls';
 export { default as TrackList } from './track-list';
+export { default as FileItem } from './file-item';
